requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.3
openai==1.64.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
termcolor==2.5.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
xxhash==3.5.0
yarl==1.18.3
